rename: enum 타입 파일 constant 패키지로 이동
