+++ src/main/java/racingcar/constant/MagicNumber.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.constant;
 
 public enum MagicNumber {
     RANDOM_MIN(0),
